Update mdev-mount.sh to avoid kernel crash
The mdev-mount.sh script and busybox mount boldy just try every
filesystemtype it knows on any partition it is called for.  On
recent kernel this may lead to an oops on swap partitions or
extended partitions.
The real probel is in the busybox mount that tries to mount the
partition. It could be much more clever by using the blkid code
(also part of busybox) to find out if the partition has a valif
FSTYPE.
As the blkid command is available it is now used in mdev-mount.sh
to exit when a sawp partition is found or when it cannot determine
a FSTYPE.
NOTE: do not use the found FSTYP in a mount -t $FSTYPE call as
it seems that this does not work (another bug in busybox???)