busybox/mdev-mount.sh: Use LABEL name for mounting
If fstab has nothing to say, use the LABEL as a mount point name. This is
what Ubuntu usually does.
It would be possible to use the UUID as an alternative too, but I don't
consider "b11d829e-6a26-472f-8683-0c8d999847d0" a nice name.
If you want your stick to be mounted as /media/usb again, either label it
"usb" using tune2fs or so, or don't label it at all.

Yes, this also means you can mount a USB disk as /media/hdd by simply
labeling it "hdd".