linux-dreambox: disable UBI for dm800se
The kernel had apparently gotten to big to boot.
As the dm800se does not use UBI anyway, the
quickest fix is to disable it again.

Not a clue why the dm500hd with an even bigger
kernel does not seem to have this problem though.
For now assume the dm800se is the only box to
suffer from this issue.