dreambox-secondstage: generate flash binary on the build system
No need to do this runtime on the target, abusing
mtd_debug to find the flash parameters, we have all
the information we need on the build host.

As a result, we no longer need the 'buildimage' binary
on the target, saving a bit of image size.

Also, keep the secondstage binary in /tmp, no need to
keep it around after we're done; saving a bit more
rootfs space.

Use 'flash_erase' to erase the mtd partition,
that's what it's for.

And finally, there is no need to sync the filesystems
after nandwrite.