Use python optimization by default (reverts OE-core 165ed464bbb9bf985dde9d8c15d000809901fff6)
In commit "python: remove the default optimization" the default was set
back to not produce .pyo files. This upsets our system as we've been
depending on .pyo files being the default for too long.

Since the OE-core change is "right" - the optimization doesn't do much
anyway and produces only confusion - this commit is to be reverted some
day. For now, just produce an image that works.