openssl: move to openssl 1.0.1
Lets give the past a rest and move to the future.

Just removing the openssl 0.9.8 tree (e.g. libssl and libcrypto).
This may break some packages in the 3rd-party feeds:

   mgcamd hypercam scam newcs vizcam.
   transmission (but there is version in the openpli-mips32el feed.
   the valihdflex packages: why does a bitratecalc tool need libcrypto?

For those who need these packages: try find a newer version or convince
the authors to rebuild against openssl-1.
Or think of some other trick...