busybox-1.21.1/mdev-mount.sh: Remove mountdir when mount failed
If mounting failed, remove the mount point. This prevents ending up
with an empty "/media/hdd" directory if mounting the harddisk failed,
which would also result in mounting it as /media/sda1 or so later on
because the /media/hdd would already be there.

(Note: Maybe it would be better to put the script into a separate
package, so we don't need to upgrade BB just for one config file)