vsftpd: Fix login with blank password
man passwd:
If the encrypted password, whether in /etc/passwd or in /etc/shadow, is an empty string, login is allowed without even asking for a password

Implement this in vsftpd. We didn't really care about the "s" anyway.

This solves not being able to login with an empty password.