volatile-media: Also load "fuse" module when needed
Usually, fuse will be in the kernel. If it's not, this will cause filesystems
that need it fail to mount and the module won't load until it's asked for.
Load the fuse module early on, allowing fuse to be built as a module.
Since volatile-media exists only to help mdev, this is as good a place as any
to load the fuse module. It has to happen before mdev loads, so it cannot wait
until modutils.